refactor(api): move controller create method
to src/team
